requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.21.0
line-bot-sdk==1.19.0
discord.py==2.0.0
requests==2.27.1
gunicorn==20.1.0